requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
Pillow==10.0.0
opencv-python==4.8.0.76
tqdm==4.66.1
flask==2.3.3
grad-cam==1.4.8
jupyter==1.0.0
ipykernel==6.25.1
torch~=2.8.0
torchvision~=0.23.0